backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
livekit-api==1.1.0
redis[hiredis]==5.2.1
pydantic==2.10.4
